+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   removing the reasoning part afterwards, to get the final answer.
 - Added `generative_type` to the output dictionaries, which can currently be either
   'base', 'instruction-tuned' or 'reasoning'. This is now used in the leaderboards.
+- Added `merge` to the output dictionaries, on whether the model is the result of a
+  merge with other models.
 - Added the summarisation dataset
   [personal-sum](https://github.com/SmartmediaAI/PersonalSum). It has been split into
   121 / 64 / 256 samples for train / validation / test, respectively, and is set to